+++ src/scss/components/BookList.scss
@@ -1,0 +1,12 @@
+.book-list {
+  width: 100%;
+  max-width: 450px;
+  padding: 10px;
+  list-style: none;
+
+  > li {
+    padding: 10px;
+    border-bottom: 1px solid #ddd;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  }
+}
