+++ src/scss/components/BookSearch.scss
@@ -1,0 +1,35 @@
+.search-section {
+  width: 100%;
+  max-width: 500px;
+  margin: 0 auto;
+
+  h2 {
+    font-size: 24px;
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
+
+  &__input-container {
+    display: flex;
+    gap: 10px;
+    margin-bottom: 20px;
+
+    input {
+      flex: 1;
+      padding: 10px;
+      border-radius: 5px;
+    }
+
+    button {
+      padding: 10px 20px;
+      border-radius: 5px;
+      background-color: #007bff;
+      color: #fff;
+    }
+  }
+
+  &__loading {
+    text-align: center;
+    padding: 20px;
+  }
+}
