+++ src/scss/components/Bookshelf.scss
@@ -1,0 +1,11 @@
+.bookshelf {
+  width: 100%;
+  max-width: 500px;
+  margin: 0 auto;
+
+  h2 {
+    font-size: 24px;
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
+}
