+++ src/tsx/components/BookSearch.tsx
@@ -1,12 +1,13 @@
 import { useState } from "react";
-import { Search, Copy } from "lucide-react";
+import { Search } from "lucide-react";
 import { GoogleBook } from "../../types/google_book";
+import "../../scss/components/BookSearch.scss";
+import BookList from "./BookList";
 
 const BookSearch = () => {
   const [query, setQuery] = useState("");
   const [results, setResults] = useState<GoogleBook[]>([]);
   const [isSearching, setIsSearching] = useState(false);
-  const [copiedId, setCopiedId] = useState<string | null>(null);
 
   const searchBooks = async () => {
     if (!query.trim()) return;
@@ -28,82 +29,29 @@ const BookSearch = () => {
     }
   };
 
-  const copyToClipboard = async (id: string) => {
-    try {
-      await navigator.clipboard.writeText(id);
-      setCopiedId(id);
-      setTimeout(() => setCopiedId(null), 2000);
-    } catch (err) {
-      console.error("Failed to copy:", err);
-    }
-  };
-
   return (
-    <div className="w-full max-w-2xl mx-auto">
-      <div className="flex gap-2 mb-6">
+    <section className="search-section">
+      <h2>書籍検索</h2>
+
+      <div className="search-section__input-container">
         <input
           type="text"
           value={query}
           onChange={(e) => setQuery(e.target.value)}
           onKeyDown={(e) => e.key === "Enter" && searchBooks()}
           placeholder="書籍名や著者名で検索..."
-          className="flex-1 p-2 border rounded-lg"
         />
-        <button
-          onClick={searchBooks}
-          disabled={isSearching}
-          className="px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 disabled:bg-blue-300"
-        >
-          <Search size={20} />
+        <button onClick={searchBooks} disabled={isSearching}>
+          <Search size={15} />
         </button>
       </div>
 
       {isSearching ? (
-        <div className="text-center py-4">検索中...</div>
+        <div className="search-section__loading">検索中...</div>
       ) : (
-        <div className="space-y-4">
-          {results.map((book) => (
-            <div key={book.id} className="flex gap-4 p-4 border rounded-lg">
-              {book.volumeInfo.imageLinks?.thumbnail && (
-                <img
-                  src={book.volumeInfo.imageLinks.thumbnail}
-                  alt={book.volumeInfo.title}
-                  className="w-24 h-32 object-cover"
-                />
-              )}
-              <div className="flex-1">
-                <h3 className="font-semibold">{book.volumeInfo.title}</h3>
-                <p className="text-sm text-gray-600">
-                  {book.volumeInfo.authors?.join(", ") || "著者不明"}
-                </p>
-                {book.volumeInfo.categories && (
-                  <p className="text-sm text-blue-600 mt-1">
-                    {book.volumeInfo.categories.join(", ")}
-                  </p>
-                )}
-                <div className="mt-2 flex items-center gap-2">
-                  <code className="bg-gray-100 px-2 py-1 rounded text-sm">
-                    {book.id}
-                  </code>
-                  <button
-                    onClick={() => copyToClipboard(book.id)}
-                    className="p-1 hover:bg-gray-100 rounded"
-                    title="IDをコピー"
-                  >
-                    <Copy size={16} />
-                  </button>
-                  {copiedId === book.id && (
-                    <span className="text-sm text-green-600">
-                      コピーしました
-                    </span>
-                  )}
-                </div>
-              </div>
-            </div>
-          ))}
-        </div>
+        <BookList books={results} />
       )}
-    </div>
+    </section>
   );
 };
